[Browser] Handle optional active_tab
The active_tab attribute can be Union(None, Tab) on initialization, but generally
the browser will always load a tab on start, so practically the type should be
Tab. To handle the narrowing, switch to using a getter/setter via @property and
only access active_tabs through the getter method.